requirements.txt
# ─── Core ML / Deep Learning ──────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=1.7.0

# ─── Data Processing ─────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# ─── Visualization ───────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ─── API Backend ─────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# ─── Utilities ───────────────────────────────────────────────────────────────
joblib>=1.3.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# ─── Web UI ──────────────────────────────────────────────────────────────────
streamlit>=1.25.0

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0
